continue reformating and removing of duplicate code
