+++ src/multilabel_classification/feature_network.py
@@ -1,14 +1,16 @@
 import torch
 import torch.nn as nn
-import torch.nn.functional as F
 import torch.optim as optim
 import numpy as np
 import os
+
 from src.multilabel_classification.utils import (
     compute_f1_score, PATH_MODELS_ML_BIB, PATH_BIBTEX,
-    load_training_set_bibtex, load_test_set_bibtex
+    load_training_set_bibtex, load_test_set_bibtex, train_for_num_epochs
 )
-from src.visualization_utils import plot_results
+from src.model.base_model import BaseModel
+from src.multilabel_classification.model.feature_mlp import FeatureMLP
+
 
 # Parameters to reproduce the baseline results of the SPEN paper
 params_baseline = {'epochs': 20, 'optim': 'adam', 'lr': 1e-3,
@@ -17,48 +19,15 @@ params_baseline = {'epochs': 20, 'optim': 'adam', 'lr': 1e-3,
 params_feature_extraction = {'epochs': 10, 'optim': 'adam', 'lr': 1e-3,
                              'momentum': 0, 'scheduler': 20, 'weight_decay': 0}
 
-FILE_FEATURE_NETWORK = "bibtex_feature_network.pth"
+FILE_FEATURE_NETWORK = "feature_network.pth"
+FILE_RESULTS_FEAT_NETWORK = "feature_network.pkl"
 PATH_FEATURE_NETWORK = os.path.join(PATH_MODELS_ML_BIB, FILE_FEATURE_NETWORK)
 
 
-class FeatureMLP(nn.Module):
+class FeatureNetwork(BaseModel):
 
-    def __init__(self, label_dim, input_dim, only_feature_extraction=False, n_hidden_units=150):
-        """
-        MLP to make a mapping from x -> F(x)
-        where F(x) is a feature representation of the inputs
-        2 layer network with sigmoid ending to predict
-        independently for each x_i its label y_i
-        n_hidden_units=150 in SPEN/INFNET papers for bibtex/Bookmarks
-        n_hidden_units=250 in SPEN/INFNET papers for Delicious
-        using Adam with lr=0.001 as the INFNET paper
-
-        Parameters:
-        ---------------
-        only_feature_extraction: bool
-            once the network is trained, we just use it until the second layer
-            for feature extraction of the inputs.
-        """
-        super().__init__()
-
-        self.only_feature_extraction = only_feature_extraction
-        self.n_hidden_units = n_hidden_units
-
-        self.fc1 = nn.Linear(input_dim, n_hidden_units)
-        self.fc2 = nn.Linear(n_hidden_units, n_hidden_units)
-        self.fc3 = nn.Linear(n_hidden_units, label_dim)
-
-    def forward(self, x):
-        x = F.relu(self.fc1(x))
-        x = F.relu(self.fc2(x))
-        if not self.only_feature_extraction:
-            x = torch.sigmoid(self.fc3(x))
-        return x
-
-
-class FeatureNetwork:
-
-    def __init__(self, use_cuda, lr=1e-3, momentum=0, optimizer='adam', weight_decay=0, input_dim=1836, label_dim=159):
+    def __init__(self, lr=1e-3, momentum=0, optimizer='adam',
+                 weight_decay=0, input_dim=1836, label_dim=159):
         """
         Model to make a word embedding
         from x --> F(x) for SPEN/INFNET models
@@ -66,8 +35,8 @@ class FeatureNetwork:
         obtained by a vanilla MLP using independent-label
         cross entropy
         """
-        self.device = torch.device("cuda" if use_cuda else "cpu")
-        self.model = FeatureMLP(label_dim, input_dim).to(self.device)
+        super().__init__(FeatureMLP(label_dim, input_dim))
+
         if optimizer == 'adam':
             self.optimizer = optim.Adam(self.model.parameters(), lr=lr, weight_decay=weight_decay)
         else:
@@ -78,7 +47,7 @@ class FeatureNetwork:
         # Our final loss is the sum over all our losses
         self.loss_fn = nn.BCELoss(reduction='sum')
 
-    def train(self, loader, ep):
+    def train(self, loader):
 
         self.model.train()
 
@@ -101,8 +70,8 @@ class FeatureNetwork:
             self.optimizer.step()
 
             if batch_idx % 10 == 0:
-                print('\rTraining Epoch {} [{} / {} ({:.0f}%)]: Avg_Loss = {:.5f}'
-                      ''.format(ep, t_size, n_train, 100 * t_size / n_train, t_loss / t_size),
+                print('\rTraining: [{} / {} ({:.0f}%)]: Avg_Loss = {:.5f}'
+                      ''.format(t_size, n_train, 100 * t_size / n_train, t_loss / t_size),
                       end='')
 
         t_loss /= t_size
@@ -138,8 +107,10 @@ class FeatureNetwork:
         mean_f1 = np.mean(mean_f1)
         loss /= t_size
         print('Avg_Loss = {:.2f}; F1_Score = {:.2f}'.format(loss.item(), 100 * mean_f1))
-
         return loss.item(), mean_f1
+
+    def test(self, loader):
+        return self.valid(loader)
 
 
 def run_the_model(do_feature_extraction: bool, path_data: str, path_save: str, use_cuda: bool):
@@ -152,30 +123,23 @@ def run_the_model(do_feature_extraction: bool, path_data: str, path_save: str, u
     else:
         params = params_baseline
 
-    f_net = FeatureNetwork(use_cuda, lr=params['lr'], momentum=params['momentum'],
+    f_net = FeatureNetwork(lr=params['lr'], momentum=params['momentum'],
                            optimizer=params['optim'], weight_decay=params['weight_decay'])
 
     results = {'name': 'MLP_Baseline', 'loss_train': [], 'loss_valid': [], 'f1_valid': []}
 
     scheduler = torch.optim.lr_scheduler.StepLR(f_net.optimizer, step_size=params['scheduler'], gamma=0.1)
 
-    best_val_found = 0
-    # Train for 10 epochs as the INFNET paper
-    for epoch in range(params['epochs']):
-        loss_train = f_net.train(train_loader, epoch)
-        loss_valid, f1_valid = f_net.valid(valid_loader)
-        scheduler.step()
-
-        results['loss_train'].append(loss_train)
-        results['loss_valid'].append(loss_valid)
-        results['f1_valid'].append(f1_valid)
-        if f1_valid > best_val_found:
-            best_val_found = f1_valid
-            print('--- Saving model at F1 = {:.2f} ---'.format(100 * best_val_found))
-            torch.save(f_net.model.state_dict(), os.path.join(path_save, FILE_FEATURE_NETWORK))
-
-    # Plot results and save the model
-    plot_results(results, iou=False)
+    train_for_num_epochs(
+        f_net,
+        train_loader,
+        valid_loader,
+        os.path.join(path_save, FILE_FEATURE_NETWORK),
+        os.path.join(path_save, FILE_RESULTS_FEAT_NETWORK),
+        params['epochs'],
+        results,
+        scheduler
+    )
 
 
 def run_test_set(path_data: str, path_save: str, use_cuda: bool):
